create index with replica only if enough nodes available
Reviewed-on: http://review.couchbase.org/119333
Tested-by: Girish Benakappa <girish.benakappa@couchbase.com>

+++ pytests/tuqquery/n1ql_fts_integration.py
@@ -526,8 +526,10 @@ class N1qlFTSIntegrationTest(QueryTests):
     def test_clusterops_fts_node_failover(self):
         self._update_replica_for_fts_index(self.fts_indexes['index1'], 1)
         self.sleep(30)
-        self.run_cbq_query("DROP index `test_bucket`.`#primary`")
-        self.run_cbq_query("CREATE PRIMARY INDEX `#primary` ON `test_bucket` with {'num_replica': 1}")
+        index_nodes = self.get_nodes_from_services_map(service_type="index", get_all_nodes=True)
+        if len(index_nodes) > 1:
+            self.run_cbq_query("DROP index `test_bucket`.`#primary`")
+            self.run_cbq_query("CREATE PRIMARY INDEX `#primary` ON `test_bucket` with {'num_replica': 1}")
         services_string = self.input.param("services_init", '')
         services = services_string.split("-")
         new_services = []
